[#1317] Address review comments
