+++ integration/hibernate-4.3/src/main/java/com/blazebit/persistence/integration/hibernate/Hibernate43Access.java
@@ -20,6 +20,7 @@ import com.blazebit.apt.service.ServiceProvider;
 import com.blazebit.persistence.integration.hibernate.base.HibernateAccess;
 import com.blazebit.persistence.integration.hibernate.base.HibernateExtendedQuerySupport;
 import com.blazebit.persistence.integration.hibernate.base.HibernateReturningResult;
+import com.blazebit.persistence.integration.hibernate.base.ScrollableResultsIterator;
 import com.blazebit.persistence.spi.DbmsDialect;
 import com.blazebit.reflection.ReflectionUtils;
 import org.hibernate.HibernateException;
@@ -331,41 +332,6 @@ public class Hibernate43Access implements HibernateAccess {
     @Override
     public ParameterTranslations createParameterTranslations(List<ParameterSpecification> queryParameterSpecifications) {
         return new ParameterTranslationsImpl(queryParameterSpecifications);
-    }
-
-    /**
-     * Iterator over {@code ScrollableResults}.
-     *
-     * @author Jan-Willem Gmelig Meyling
-     * @since 1.6.2
-     */
-    private static class ScrollableResultsIterator implements Iterator<Object> {
-        private final ScrollableResults scrollableResults;
-
-        public ScrollableResultsIterator(ScrollableResults scrollableResults) {
-            this.scrollableResults = scrollableResults;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return scrollableResults.next();
-        }
-
-        @Override
-        public Object next() {
-            Object[] next = scrollableResults.get();
-            if (next.length == 1) {
-                return next[0];
-            } else {
-                return next;
-            }
-        }
-
-        @Override
-        public void remove() {
-            throw new UnsupportedOperationException("remove");
-        }
-
     }
 
 }
