+++ integration/hibernate-5.3/src/main/java/com/blazebit/persistence/integration/hibernate/Hibernate53Access.java
@@ -19,6 +19,7 @@ package com.blazebit.persistence.integration.hibernate;
 import com.blazebit.apt.service.ServiceProvider;
 import com.blazebit.persistence.integration.hibernate.base.HibernateAccess;
 import com.blazebit.persistence.integration.hibernate.base.HibernateReturningResult;
+import com.blazebit.persistence.integration.hibernate.base.ScrollableResultsIterator;
 import com.blazebit.persistence.spi.DbmsDialect;
 import org.hibernate.HibernateException;
 import org.hibernate.LockOptions;
@@ -259,32 +260,4 @@ public class Hibernate53Access implements HibernateAccess {
         }
     }
 
-    /**
-     * Iterator over {@code ScrollableResultsImplementor}.
-     *
-     * @author Jan-Willem Gmelig Meyling
-     * @since 1.6.2
-     */
-    private static class ScrollableResultsIterator implements Iterator<Object> {
-        private final ScrollableResultsImplementor scrollableResults;
-
-        public ScrollableResultsIterator(ScrollableResultsImplementor scrollableResults) {
-            this.scrollableResults = scrollableResults;
-        }
-
-        @Override
-        public boolean hasNext() {
-            return !scrollableResults.isClosed() && scrollableResults.next();
-        }
-
-        @Override
-        public Object next() {
-            Object[] next = scrollableResults.get();
-            if (next.length == 1) {
-                return next[0];
-            } else {
-                return next;
-            }
-        }
-    }
 }
